fix(sparse): Validate i32 indices in usize domain
gather/scatter and the CSR/SELL-p column validators bounded i32 indices
against `len as i32`, which truncates for 2^31-element views or column
counts: valid indices were rejected (negative bound) or misreported.
Compare in usize after the sign check instead — exact for every length.

Item: HS-SPARSE-SAFETY-2026-08-27

+++ crates/hermes-simd-core/src/sparse/types.rs
@@ -458,7 +458,10 @@ impl<T, V: AsRef<[T]>, I: AsRef<[i32]>> SparseValidate for CsrMatrix<T, V, I> {
             }
         }
         for &col in col_indices {
-            if col < 0 || col >= self.ncols as i32 {
+            // Compare in `usize` after the sign check: an `i32` bound
+            // (`ncols as i32`) would truncate for 2^31 or more columns and
+            // misreport valid indices.
+            if col < 0 || col as usize >= self.ncols {
                 return Err(crate::SimdError::IndexOutOfBounds);
             }
         }
@@ -501,7 +504,10 @@ impl<T, const C: usize, V: AsRef<[T]>, I: AsRef<[i32]>> SparseValidate for SellP
             }
         }
         for &col in col_indices {
-            if col < 0 || col >= self.ncols as i32 {
+            // Compare in `usize` after the sign check: an `i32` bound
+            // (`ncols as i32`) would truncate for 2^31 or more columns and
+            // misreport valid indices.
+            if col < 0 || col as usize >= self.ncols {
                 return Err(crate::SimdError::IndexOutOfBounds);
             }
         }
